Move sudoku utils to package
