+++ apps/website/src/components/Sudoku/Display/Clues/Logic.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
-  import { cellSize, symbolsMap } from '$constants';
+  import { symbolsMap } from '$constants';
   import type { Dimensions, Logic } from '@octopuzzles/models';
+  import { CELL_SIZE } from '@octopuzzles/models';
   import { arrayfrom0ToN } from '@octopuzzles/utils';
 
   export let logic: Logic;
@@ -14,8 +15,8 @@
         {#each arrayfrom0ToN(dimensions.rows - (dimensions.margins ? dimensions.margins.top + dimensions.margins.bottom : 0)) as _, rowIndex}
           <svelte:component
             this={symbolsMap.Diagonal}
-            x={((dimensions.margins?.left ?? 0) + rowIndex) * cellSize}
-            y={(dimensions.rows - (dimensions.margins?.bottom ?? 0) - rowIndex - 1) * cellSize}
+            x={((dimensions.margins?.left ?? 0) + rowIndex) * CELL_SIZE}
+            y={(dimensions.rows - (dimensions.margins?.bottom ?? 0) - rowIndex - 1) * CELL_SIZE}
             rotation={'East'}
             color={'Black'}
           />
@@ -24,8 +25,8 @@
         {#each arrayfrom0ToN(dimensions.rows - (dimensions.margins ? dimensions.margins.top + dimensions.margins.bottom : 0)) as _, rowIndex}
           <svelte:component
             this={symbolsMap.Diagonal}
-            x={((dimensions.margins?.left ?? 0) + rowIndex) * cellSize}
-            y={((dimensions.margins?.top ?? 0) + rowIndex) * cellSize}
+            x={((dimensions.margins?.left ?? 0) + rowIndex) * CELL_SIZE}
+            y={((dimensions.margins?.top ?? 0) + rowIndex) * CELL_SIZE}
             rotation={'North'}
             color={'Black'}
           />
@@ -33,20 +34,20 @@
       {:else if flag === 'Indexed159' && dimensions.rows === 9 && dimensions.columns === 9}
         {#each arrayfrom0ToN(dimensions.rows - (dimensions.margins ? dimensions.margins.top + dimensions.margins.bottom : 0)) as _, rowIndex}
           <rect
-            x={cellSize * (dimensions.margins?.left ?? 0)}
-            y={cellSize * ((dimensions.margins?.top ?? 0) + rowIndex)}
+            x={CELL_SIZE * (dimensions.margins?.left ?? 0)}
+            y={CELL_SIZE * ((dimensions.margins?.top ?? 0) + rowIndex)}
             class="fill-current text-red w-cell h-cell opacity-60"
             vector-effect="non-scaling-size"
           />
           <rect
-            x={cellSize * ((dimensions.margins?.left ?? 0) + 4)}
-            y={cellSize * ((dimensions.margins?.top ?? 0) + rowIndex)}
+            x={CELL_SIZE * ((dimensions.margins?.left ?? 0) + 4)}
+            y={CELL_SIZE * ((dimensions.margins?.top ?? 0) + rowIndex)}
             class="fill-current text-red w-cell h-cell opacity-60"
             vector-effect="non-scaling-size"
           />
           <rect
-            x={cellSize * ((dimensions.margins?.left ?? 0) + 8)}
-            y={cellSize * ((dimensions.margins?.top ?? 0) + rowIndex)}
+            x={CELL_SIZE * ((dimensions.margins?.left ?? 0) + 8)}
+            y={CELL_SIZE * ((dimensions.margins?.top ?? 0) + rowIndex)}
             class="fill-current text-red w-cell h-cell opacity-60"
             vector-effect="non-scaling-size"
           />
