+++ apps/website/src/components/Sudoku/Display/Clues/Notes.svelte
@@ -1,15 +1,14 @@
 <script lang="ts">
-  import { cellSize } from '$constants';
-  import type { Position } from '@octopuzzles/models';
-  import type { Notes } from '@octopuzzles/models';
+  import { CELL_SIZE } from '@octopuzzles/models';
+  import type { Position, Notes } from '@octopuzzles/models';
 
   export let notes: Notes | undefined;
   export let onClickNote: ((note: string, position: Position) => void) | undefined = undefined;
 
   function createPolygonPoints(row: number, column: number): string {
-    let firstPoint = `${cellSize * column + 0.8 * cellSize},${cellSize * row}`;
-    let secondPoint = `${cellSize * (column + 1)},${cellSize * row}`;
-    let thirdPoint = `${cellSize * (column + 1)},${cellSize * row + 0.2 * cellSize}`;
+    let firstPoint = `${CELL_SIZE * column + 0.8 * CELL_SIZE},${CELL_SIZE * row}`;
+    let secondPoint = `${CELL_SIZE * (column + 1)},${CELL_SIZE * row}`;
+    let thirdPoint = `${CELL_SIZE * (column + 1)},${CELL_SIZE * row + 0.2 * CELL_SIZE}`;
 
     return `${firstPoint} ${secondPoint} ${thirdPoint}`;
   }
