+++ apps/website/src/components/Sudoku/Display/Clues/Numbers.svelte
@@ -1,7 +1,6 @@
 <script lang="ts">
-  import { cellSize } from '$constants';
-  import type { Dimensions, Givens } from '@octopuzzles/models';
-  import type { GameValues } from '@octopuzzles/models';
+  import { CELL_SIZE } from '@octopuzzles/models';
+  import type { Dimensions, Givens, GameValues } from '@octopuzzles/models';
   import { arrayfrom0ToN } from '@octopuzzles/utils';
   import classNames from 'classnames';
 
@@ -30,8 +29,8 @@
       {@const val = givens[row][column] || values?.[row]?.[column]}
       {#if val && val.length > 0}
         <text
-          x={cellSize * (column + 0.5)}
-          y={cellSize * (row + 0.55)}
+          x={CELL_SIZE * (column + 0.5)}
+          y={CELL_SIZE * (row + 0.55)}
           dominant-baseline="middle"
           class={classNames(
             'fill-current text-4xl textanchor-middle',
