+++ packages/sudoku-utils/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@octopuzzles/utils",
+  "name": "@octopuzzles/sudoku-utils",
   "version": "0.0.0",
   "private": true,
   "main": "./dist/index.js",
@@ -13,6 +13,9 @@
     "clean": "rm -rf dist",
     "lint": "eslint \"src/**/*.ts*\"",
     "test:unit": "vitest run"
+  },
+  "dependencies": {
+    "@octopuzzles/models": "*"
   },
   "devDependencies": {
     "eslint": "^7.32.0",
